requirements.txt
Flask==1.1.2
futures==3.1.1
gunicorn==20.0.4
newspaper3k==0.2.8
nltk==3.4.1
pymongo==3.8.0
requests==2.19.1
Flask-Cors==3.0.8
Flask-Caching==1.8.0